before authorization

+++ routes/ngos.js
@@ -2,14 +2,13 @@ const express = require('express')
 const catchAsync = require("../views/utils/catchAsync");
 const con = require("../models/ngo");
 const router = express.Router()
-const { v4: UUID } = require('uuid');
+const {v4: UUID} = require('uuid');
 const {Strategy: LocalStrategy} = require("passport-local");
-
-
+const {CLIENT_ODBC} = require("mysql/lib/protocol/constants/client");
 
 
 router.get('/', catchAsync(async (req, res) => {
-
+    // console.log(req.session.user)
     const sql = `SELECT * FROM ngoSchema`
     await con.query(sql, (error, result) => {
         if (error) console.log('\n\n\n\n\nERROR!!!!\n\n\n\n\n' + error)
@@ -35,44 +34,64 @@ router.post('/', catchAsync(async (req, res) => {
     const description = req.body.description
     const location = req.body.location
     const id = UUID()
+    const userID = await req.session.user.userID
 
-    const queryAdd = `insert into ngoSchema (id, title, description, location) values (?, ?, ?, ?)`
-    await con.query(queryAdd, [id, title, description, location], (error, result) => {
+
+
+
+
+
+
+    const queryAdd = `insert into ngoSchema (id, title, description, location, userID) values (?, ?, ?, ?, ?)`
+    await con.query(queryAdd, [id, title, description, location, userID], (error, result) => {
         if (error) console.log('\n\n\n\n\nERROR!!!!\n\n\n\n\n' + error)
         req.flash('success', 'Added new ngo')
         res.redirect(`/ngos/${id}`)
     })
-
 }))
 
 router.get('/:id', catchAsync(async (req, res) => {
     const ngoID = await req.params.id
+    const userID = await req.session.user.userID
     const queryShow = `SELECT * from ngoSchema where id = ?`
-    const queryFeedback = `SELECT * from feedbackSchema where ngoid = ?`
+    const queryFeedback = `SELECT * from feedbackSchema where ngoID = ?`
 
-    await con.query(queryShow, [ngoID], (error, nResult) => {
-        if (error) {
-            return res.status(500).send(error);
-        }
-        con.query(queryFeedback, [ngoID], (error, fResult) => {
-            if (error) {
-                return res.status(500).send(error);
+    const adminDetails = `SELECT * from userSchema where userID = ?`
+    con.query(adminDetails, [userID], (err, aResult) => {
+            if (err) {
+                return res.status(500).send(err);
+            } else {
+                con.query(queryShow, [ngoID], (error, nResult) => {
+                    if (error) {
+                        return res.status(500).send(error);
+                    }
+                    con.query(queryFeedback, [ngoID], (error, fResult) => {
+                        if (error) {
+                            return res.status(500).send(error);
+                        }
+                        res.render('ngos/show', {ngo: nResult[0], feedback: fResult, admin: aResult[0]})
+                    })
+                })
             }
-            res.render('ngos/show', {ngo: nResult[0], feedback: fResult})
-        })
-    })
+        }
+    )
 }))
 
 router.get('/:id/edit', catchAsync(async (req, res) => {
-    const id = req.params.id
-    const queryShow = `SELECT * from ngoSchema where id = ?`
+    if (req.session.user) {
+        const id = req.params.id
+        const queryShow = `SELECT * from ngoSchema where id = ?`
 
-    await con.query(queryShow, [id], (error, result) => {
-        if (error) {
-            return res.status(500).send(error)
-        }
-        res.render('ngos/edit', {ngo: result[0]})
-    })
+        await con.query(queryShow, [id], (error, result) => {
+            if (error) {
+                return res.status(500).send(error)
+            }
+            res.render('ngos/edit', {ngo: result[0]})
+        })
+    } else {
+        req.flash('You must be logged in')
+        res.redirect('/login');
+    }
 }))
 
 router.post('/:id', catchAsync(async (req, res) => {
@@ -93,19 +112,23 @@ router.post('/:id', catchAsync(async (req, res) => {
 
 
 router.delete('/:id', catchAsync(async (req, res) => {
-    const id = req.params.id
-    const queryDelete = `DELETE FROM ngoSchema WHERE id = ?`
-    const queryDeleteFeedback = `DELETE FROM feedbackSchema WHERE ngoID = ?`
 
-    await con.query(queryDeleteFeedback, [id], (error, result) => {
-        if (error) console.log('\n\n\n\n\nERROR!!!!\n' + error)
-        con.query(queryDelete, [id], (error, result) => {
-            res.redirect('/ngos')
+    if (req.session.user) {
+        const id = req.params.id
+        const queryDelete = `DELETE FROM ngoSchema WHERE id = ?`
+        const queryDeleteFeedback = `DELETE FROM feedbackSchema WHERE ngoID = ?`
+
+        await con.query(queryDeleteFeedback, [id], (error, result) => {
+            if (error) console.log('\n\n\n\n\nERROR!!!!\n' + error)
+            con.query(queryDelete, [id], (error, result) => {
+                res.redirect('/ngos')
+            })
         })
-    })
+    } else {
+        req.flash('You must be logged in')
+        res.redirect('/login');
+    }
 }))
 
 module.exports = router
-
-
